Crate scoped_pool

scoped-pool

A flexible thread pool providing scoped threads.

Structs

Pool
A thread-pool providing scoped and unscoped threads.

Scope
An execution scope, represents a set of jobs running on a Pool.

ThreadConfig
Thread configuration. Provides detailed control over the properties and behavior of new threads.

WaitGroup
A synchronization primitive for awaiting a set of actions.